RavenDB-20180: move files
